requirements.txt
aiohttp==3.8.3
aiosignal==1.3.1
async-timeout==4.0.2
attrs==22.1.0
charset-normalizer==2.1.1
discord.py==2.1.0
frozenlist==1.3.3
idna==3.4
multidict==6.0.3
yarl==1.8.2
requests==2.26.0
json==0.9.7
discord==1.0.1
os==1.11.2
appdirs==1.4.4
appnope==0.1.2
distlib==0.3.3
filelock==3.0.12
importlib-metadata==4.8.1
typing-extensions==4.1.1
zipp==3.6.0
regex==2022.2.22
beautifulsoup4==4.10.0
python-docx==0.9.3
python-pptx==0.7.0
google-api-python-client  
google-auth-oauthlib==0.4.6  
youtube-transcript-api==0.4.1  
